requirements.txt
asgiref==3.2.10
Django==3.0.8
django-address==0.2.4
django-debug-toolbar==2.2
django-money==1.1
django-phone-field==1.8.1
Pillow==7.2.0
py-moneyed==0.8.0
pytz==2020.1
sqlparse==0.3.1
